app_bot/requirements.txt
git+ssh://git@github.com/dumaevrinat/lung_diseases.git@main#egg=lungs_ml&subdirectory=lungs_ml
pyTelegramBotAPI==3.7.6
numpy==1.19.2
Pillow==8.2.0
